Delegate some logic to Dynamics backend
