+++ app/services/bookings/gitis/backends.rb
@@ -1,0 +1,9 @@
+module Bookings
+  module Gitis
+    module Backends
+      def self.load(*args)
+        Dynamics.new(*args)
+      end
+    end
+  end
+end
